refactor: Move tests to integration tests folder
